chore(client): :truck: Create separate package for client
The Intent is to maximize modularization
